feat: remove unused driver and application files, consolidate project structure
