+++ Cargo.toml
@@ -1,17 +1,31 @@
-[workspace]
-members = ["app", "drivers"]
-resolver = "2"
+[package]
+name = "s3-display-amoled-touch-app"
+version = "0.1.0"
+edition = "2021"
 
-[profile.dev]
-# Rust debug is too slow.
-# For debug builds always builds with some optimization
-opt-level = "s"
+[dependencies]
+esp-backtrace = { version = "0.15.0", features = [
+    "esp32s3",
+    "exception-handler",
+    "panic-handler",
+    "defmt",
+] }
+mipidsi = { git = "https://github.com/almindor/mipidsi.git", branch = "master" }
+bq25896x = { git = "https://github.com/psytraxx/bq25896x-rs.git", branch = "main" }
+esp-hal = { version = "0.23.0", features = ["esp32s3", "defmt", "octal-psram"] }
+esp-alloc = { version = "0.6.0" }
+defmt = "0.3.8"
+defmt-rtt = "0.4.1"
+embedded-hal = { version = "1.0.0", features = ["defmt-03"] }
+embedded-graphics-core = { version = "0.4.0", features = ["defmt"] }
+embedded-hal-bus = { version = "0.3.0", features = ["defmt-03"] }
+libm = "0.2.11"
+slint = { version = "1.9.1", default-features = false, features = [
+    "libm",
+    "unsafe-single-threaded",
+    "compat-1-2",
+    "renderer-software",
+] }
 
-[profile.release]
-codegen-units = 1        # LLVM can perform better optimizations using a single thread
-debug = 2
-debug-assertions = false
-incremental = false
-lto = 'fat'
-opt-level = 's'
-overflow-checks = false
+[build-dependencies]
+slint-build = "1.9.1"
