+++ assets/battery.png
