--- drivers/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "s3-display-amoled-touch-drivers"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-embedded-hal = { version = "1.0.0" }
-libm = "0.2"
