--- drivers/src/lib.rs
@@ -1,7 +1,0 @@
-#![no_std]
-
-#[macro_use]
-extern crate alloc;
-
-pub mod bq25896;
-pub mod cst816s;
