+++ src/main.rs
@@ -5,7 +5,9 @@
 use alloc::boxed::Box;
 use alloc::rc::Rc;
 use alloc::string::ToString;
+use bq25896x::bq25896::{ChargeStatus, PmuSensorError, BQ25896};
 use core::time::Duration;
+use cst816s::CST816S;
 use defmt::{error, info};
 use draw_buffer::DrawBuffer;
 use embedded_hal::i2c::I2c as I2CBus;
@@ -28,8 +30,6 @@ use esp_hal::timer::timg::TimerGroup;
 use esp_hal::xtensa_lx::singleton;
 use esp_hal::{dma_buffers, main};
 use mipidsi::interface::SpiInterface;
-use s3_display_amoled_touch_drivers::bq25896::{ChargeStatus, PmuSensorError};
-use s3_display_amoled_touch_drivers::{bq25896::BQ25896, cst816s::CST816S};
 use slint::platform::software_renderer::{MinimalSoftwareWindow, RepaintBufferType, Rgb565Pixel};
 use slint::platform::{Platform, PointerEventButton};
 use slint::{LogicalPosition, PhysicalSize};
@@ -39,6 +39,7 @@ slint::include_modules!();
 
 extern crate alloc;
 
+mod cst816s;
 mod draw_buffer;
 
 pub const DISPLAY_HEIGHT: u16 = 240;
